requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Werkzeug==2.3.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
itsdangerous==2.1.2